+++ paging/src/commonMain/kotlin/org/mobilenativefoundation/store/paging5/KeyFactory.kt
@@ -1,0 +1,9 @@
+package org.mobilenativefoundation.store.paging5
+
+import org.mobilenativefoundation.store.core5.ExperimentalStoreApi
+import org.mobilenativefoundation.store.core5.StoreKey
+
+@ExperimentalStoreApi
+interface KeyFactory<Id : Any, SK : StoreKey.Single<Id>> {
+    fun createFor(id: Id): SK
+}
